Address Float64 precision loss in duration_to_end / duration_to_start
Bot review on PR #105319 flagged that the form
`Float64(current_timestamp) - Float64(last_timestamp)` rounds each operand
to ~256 ns quanta at `DateTime64(9)` near present-day epoch (~1.7e18), so
exact 1 ns deltas are silently zeroed. Subtract in `Int128` first, convert
to `Float64` once. Same change applied symmetrically to `duration_to_start`.

Adds `04271_timeseries_extrapolated_value_float64_precision` which fails
before the fix (rate `[1.5e-9]`, delta `[1.5]`) and passes after it
(`[2e-9]`, `[2]`).

+++ src/AggregateFunctions/TimeSeries/AggregateFunctionTimeseriesExtrapolatedValue.h
@@ -139,11 +139,16 @@ private:
         /// The following logic is copied from Prometheus' rate calculation
         /// https://github.com/prometheus/prometheus/blob/5e124cf4f2b9467e4ae1c679840005e727efd599/promql/functions.go#L127
         /// which is licensed under the Apache License 2.0
-        // Duration between first/last samples and boundary of range. Computed in `Float64` to avoid signed overflow on `current_timestamp - Base::window`.
-        Float64 duration_to_start =
-            static_cast<Float64>(first_timestamp) - static_cast<Float64>(current_timestamp)
-            + static_cast<Float64>(Base::window);
-        Float64 duration_to_end = static_cast<Float64>(current_timestamp) - static_cast<Float64>(last_timestamp);
+        // Duration between first/last samples and boundary of range. Subtract in `Int128` first to avoid
+        // both signed overflow on `current_timestamp - Base::window` and `Float64` precision loss when
+        // timestamps are large (e.g. `DateTime64(9)` near present-day epoch ~1.7e18).
+        Float64 duration_to_start = static_cast<Float64>(
+            static_cast<Int128>(static_cast<Int64>(first_timestamp))
+            - static_cast<Int128>(static_cast<Int64>(current_timestamp))
+            + static_cast<Int128>(static_cast<Int64>(Base::window)));
+        Float64 duration_to_end = static_cast<Float64>(
+            static_cast<Int128>(static_cast<Int64>(current_timestamp))
+            - static_cast<Int128>(static_cast<Int64>(last_timestamp)));
 
         const auto sampled_interval = time_difference;
         const Float64 average_duration_between_samples = static_cast<Float64>(sampled_interval) / static_cast<Float64>(samples_in_window.size() - 1);
